requirements.txt
# Dependencias principales
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
Pillow==10.1.0

# Autenticación y seguridad
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
python-decouple==1.0.0

# Comunicación y tareas asíncronas
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0

# Integraciones
requests==2.31.0

# Tiempo real
channels==4.0.0
daphne==4.0.0

# Documentación API
drf-yasg==1.21.7

# Desarrollo y pruebas
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2
factory-boy==3.3.0

# Validación de datos
django-phonenumber-field==7.2.0
phonenumbers==8.13.25

# Filtros y búsqueda
django-filter==23.5